+++ reddwarf/data_presenter.py
@@ -27,7 +27,7 @@ def generate_figure(
     plt.figure(figsize=(7, 5), dpi=80)
     plt.axhline(y=0, color='k', linestyle='-', linewidth=0.5)
     plt.axvline(x=0, color='k', linestyle='-', linewidth=0.5)
-    plt.gca().invert_xaxis()
+    plt.gca().invert_yaxis()
 
     # Label points with participant_id if no labels set.
     for participant_id, row in coord_dataframe.iterrows():
@@ -56,8 +56,9 @@ def generate_figure(
                 # TODO: Accomodate 2 points like Polis platform does.
                 print("Cannot create concave hull for less than 3 points. Skipping...")
                 continue
-            vertex_indices = concave_hull_indexes(np.asarray(points_df), concavity=4.0)
+            vertex_indices = concave_hull_indexes(np.asarray(points_df.loc[:, ["x", "y"]]), concavity=4.0)
             hull_points = points_df.iloc[vertex_indices, :]
+            hull_points = hull_points.loc[:, ["x", "y"]]
             polygon = patches.Polygon(
                 hull_points,
                 fill=True,