+++ reddwarf/utils/__init__.py
@@ -3,3 +3,4 @@ from reddwarf.utils.matrix import *
 from reddwarf.utils.polismath import *
 from reddwarf.utils.stats import *
 from reddwarf.utils.pca import *
+from reddwarf.utils.statements import *