+++ reddwarf/utils/pca.py
@@ -37,15 +37,30 @@ def run_pca(
     # TODO: Investigate why some numbers are a bit off here.
     #       ANSWER: Because centers are calculated on unfiltered raw matrix for some reason.
     #       means = -raw_vote_matrix.mean(axis="rows")
-    means = -pca.mean_
+    means = pca.mean_
 
     # Project participant vote data onto 2D using eigenvectors.
-    projected_data = pca.transform(imputed_matrix)
-    projected_data = pd.DataFrame(projected_data, index=imputed_matrix.index, columns=np.asarray(["x", "y"]))
+    # TODO: Determine what exactly we want to be doing here.
+
+    # (1) This is what we used to use:
+    # projected_data = pca.transform(imputed_matrix)
+
+    # (2) Perhaps we could be doing this instead: (would need cleanup to clear out NaN values)
+    # projected_data = pca.transform(vote_matrix)
+
+    # (3) This is what we'll do for now, as it reproduced Polis calculations exactly:
+    # Project data from raw, non-imputed vote_matrix.
+    # TODO: Figure out why signs are flipped here after custom projection, unlike pca.transform().
+    # Not required for regular pca.transform(), so perhaps a polismath BUG?
+    # We fix this in implementations.run_clustering().
+    projected_data = [sparsity_aware_project_ptpt(ptpt_votes, pca.components_, pca.mean_) for pid, ptpt_votes in vote_matrix.iterrows()]
+
+    projected_data = pd.DataFrame(projected_data, index=vote_matrix.index, columns=np.asarray(["x", "y"]))
     projected_data.index.name = "participant_id"
 
     return projected_data, eigenvectors, eigenvalues, means
 
+# This isn't used right now, as we're doing the scaling in the port of `sparcity_aware_project_ptpt()`.
 def scale_projected_data(
         projected_data: pd.DataFrame,
         vote_matrix: VoteMatrix
@@ -78,7 +93,7 @@ def sparsity_aware_project_ptpt(participant_votes, statement_components, stateme
     Projects a sparse vote vector into PCA space while adjusting for sparsity.
 
     Args:
-        participant_votes: List of participant votes on each statement
+        participant_votes (list): List of participant votes on each statement
         statement_components (list[list[float]]): Two lists of floats corresponding to the two principal components
         statement_means (list[float]): List of floats corresponding to the centers/means of each statement
 
@@ -88,6 +103,7 @@ def sparsity_aware_project_ptpt(participant_votes, statement_components, stateme
     statement_components = np.array(statement_components)  # Shape: (2, n_features)
     statement_means = np.array(statement_means)  # Shape: (n_features,)
 
+    # TODO: This included zerod out (moderated) statements. Should it?
     n_statements = len(participant_votes)
 
     participant_votes = np.array(participant_votes)
@@ -101,7 +117,8 @@ def sparsity_aware_project_ptpt(participant_votes, statement_components, stateme
     p1 = np.dot(x_vals, pc1_vals)
     p2 = np.dot(x_vals, pc2_vals)
 
-    n_votes = np.count_nonzero(mask)  # Non-null votes count
+    # Non-null votes count
+    n_votes = np.count_nonzero(mask)
     scale = np.sqrt(n_statements / max(n_votes, 1))
 
     projected_coord = scale * np.array([p1, p2])